requirements.txt
django
environs
djangorestframework
firebase-admin
whitenoise
drf-yasg
psycopg2-binary
supabase
django-filter
django-cors-headers
python-dotenv
python-decouple